begin van leden component en login

+++ bieb-app-productie-code/bieb-app/src/main/java/com/leesenleer/BiebApp/services/BiebService.java
@@ -4,6 +4,7 @@ import com.leesenleer.BiebApp.model.Bieb;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
-public interface BiebService extends CrudRepository<Bieb, Integer> {
-}
+public interface BiebService extends CrudRepository<Bieb, Long> {
+  }
